+++ features/claims/advocate/advocate_claim_draft_edit_submit.feature
@@ -1,18 +1,15 @@
-@javascript @webmock_allow_localhost_connect
+@javascript
 Feature: Advocate partially fills out a draft claim for a trial, then later edits and submits it
 
   Scenario: I create a claim, save it to draft and later complete it
 
     Given I am a signed in advocate
-    And There are case and fee types in place
-    And There are certification types in place
-    And There are courts, offences and expense types in place
     And I am on the 'Your claims' page
     And I click 'Start a claim'
     Then I should be on the new claim page
 
     When I select an advocate category of 'Junior alone'
-    And I select a court
+    And I select the court 'Blackfriars Crown'
     And I select a case type of 'Trial'
     And I enter a case number of 'A12345678'
     And I select an offence category
