+++ features/claims/litigator/interim_trial_claim_draft_submit.feature
@@ -1,22 +1,31 @@
-@javascript @webmock_allow_localhost_connect
-Feature: Litigator submits a claim for a Contempt case
+@javascript
+Feature: Litigator partially fills out a draft interim claim, then later edits and submits it
 
-  Scenario: I create a contempt claim, then submit it
+  Scenario: I create an interim claim, save it to draft and later complete it
 
     Given I am a signed in litigator
-    And There are supplier numbers in place
-    And There are case and fee types in place
-    And There are certification types in place
-    And There are courts, offences and expense types in place
-    And There are disbursement types in place
+    And I am allowed to submit interim claims
+    And My provider has supplier numbers
     And I am on the 'Your claims' page
+
     And I click 'Start a claim'
-    Then I should be on the litigator new claim page
+    And I select the fee scheme 'Litigator interim fee'
+    Then I should be on the litigator new interim claim page
 
     When I select the supplier number '1A222Z'
-    And I select a court
-    And I select a case type of 'Contempt'
+    And I select the court 'Blackfriars Crown'
+    And I select a case type of 'Trial'
     And I enter a case number of 'A12345678'
+
+    And I save as draft
+    Then I should see 'Draft claim saved'
+
+    Given I am later on the Your claims page
+    Then Claim 'A12345678' should be listed with a status of 'Draft'
+
+    When I click the claim 'A12345678'
+    And I edit this claim
+
     And I select the offence class 'E: Burglary'
     And I enter the case concluded date
     And I enter defendant, representation order and MAAT reference
@@ -24,14 +33,11 @@ Feature: Litigator submits a claim for a Contempt case
 
     Then I click "Continue" in the claim form
 
-    And I fill '100.25' as the fixed fee total
-    And I add a miscellaneous fee 'Costs judge application'
-    And I add a Case uplift fee with case numbers 'A12345678, A12345588'
+    And I select an interim fee type of 'Disbursement only'
     And I add a disbursement 'Computer experts' with net amount '125.40' and vat amount '30.5'
-    And I add a disbursement 'Meteorologist' with net amount '58.22' and vat amount '10.3'
-    And I add an expense 'Parking'
+    And I add another disbursement 'Meteorologist' with net amount '58.22' and vat amount '0'
 
-    And I upload 3 documents
+    And I upload 1 document
     And I check the boxes for the uploaded documents
     And I add some additional information
     And I click Submit to LAA
