+++ features/claims/litigator/litigator_contempt_claim_draft_submit.feature
@@ -1,20 +1,17 @@
-@javascript @webmock_allow_localhost_connect
-Feature: Litigator partially fills out a draft claim for a Contempt, then later edits and submits it
+@javascript
+Feature: Litigator partially fills out a draft final fee claim, then later edits and submits it
 
-  Scenario: I create a contempt claim, save it to draft and later complete it
+  Scenario: I create a final fee claim, save it to draft and later complete it
 
     Given I am a signed in litigator
-    And There are supplier numbers in place
-    And There are case and fee types in place
-    And There are certification types in place
-    And There are courts, offences and expense types in place
-    And There are disbursement types in place
+    And I am not allowed to submit interim or transfer claims
+    And My provider has supplier numbers
     And I am on the 'Your claims' page
     And I click 'Start a claim'
     Then I should be on the litigator new claim page
 
     When I select the supplier number '1A222Z'
-    And I select a court
+    And I select the court 'Blackfriars Crown'
     And I select a case type of 'Contempt'
     And I enter a case number of 'A12345678'
 
@@ -30,12 +27,15 @@ Feature: Litigator partially fills out a draft claim for a Contempt, then later 
     And I select the offence class 'E: Burglary'
     And I enter the case concluded date
     And I enter defendant, representation order and MAAT reference
+    And I add another defendant, representation order and MAAT reference
 
     Then I click "Continue" in the claim form
 
     And I fill '100.25' as the fixed fee total
     And I add a miscellaneous fee 'Costs judge application'
-    And I add a disbursement 'Meteorologist' with net amount '58.22' and vat amount '10.3'
+    And I add a Case uplift fee with case numbers 'A12345678, A12345588'
+    And I add a disbursement 'Computer experts' with net amount '125.40' and vat amount '30.5'
+    And I add another disbursement 'Meteorologist' with net amount '58.22' and vat amount '0'
     And I add an expense 'Parking'
 
     And I upload 1 document
