+++ features/page_objects/sections/common_date_section.rb
@@ -1,0 +1,6 @@
+class CommonDateSection < SitePrism::Section
+  include DateHelper
+  element :day,   'div.form-date > div.form-group-day > input'
+  element :month, 'div.form-date > div.form-group-month > input'
+  element :year,  'div.form-date > div.form-group-year > input'
+end
