+++ features/page_objects/sections/fee_section.rb
@@ -1,0 +1,5 @@
+class FeeSection < SitePrism::Section
+  element :quantity, "input.quantity"
+  element :rate, "input.rate"
+  element :add_dates, "td:nth-of-type(5) > a"
+end
